refactor 1. GameResult 가 모든 게임 결과를 상태로 가지고있으므로, 게임 결과 중 마지막을 기준으로 최종 승자 가리는것으로 역할지정
