+++ src/main/java/racingGame/model/GameResult.java
@@ -10,6 +10,26 @@ public record GameResult(List<ProgressRecord> progressRecords) {
         this.progressRecords = new ArrayList<>(progressRecords);
     }
     
+    public List<Car> findWinners() {
+        return findMaxCar(findMaxLocation());
+    }
     
+    private int findMaxLocation() {
+        int max = Integer.MIN_VALUE;
+        for(Car joinCar: progressRecords.getLast().joinCars()) {
+            max = Math.max(max, joinCar.findLocation());
+        }
+        return max;
+    }
+    
+    private List<Car> findMaxCar(int max) {
+        List<Car> winnerCars = new ArrayList<>();
+        for(Car joinCar: progressRecords.getLast().joinCars()) {
+            if(max == joinCar.findLocation()) {
+                winnerCars.add(joinCar);
+            }
+        }
+        return winnerCars;
+    }
 
 }
