+++ src/test/java/racingGame/model/ProgressRecordTest.java
@@ -30,34 +30,6 @@ class ProgressRecordTest {
     }
     
     @Test
-    void findJointWinnersTest() {
-        List<Car> joinCars = new ArrayList<>();
-        joinCars.add(new Car("pobi",5));
-        joinCars.add(new Car("crong",5));
-        joinCars.add(new Car("mo",3));
-        joinCars.add(new Car("jae",2));
-        progressRecord = new ProgressRecord(joinCars);
-        
-        List<Car> winner = List.of(new Car("pobi",5), new Car("crong",5));
-        
-        assertThat(progressRecord.findWinners()).isEqualTo(winner);
-    }
-    
-    @Test
-    void findSoleWinnersTest() {
-        List<Car> joinCars = new ArrayList<>();
-        joinCars.add(new Car("pobi",4));
-        joinCars.add(new Car("crong",5));
-        joinCars.add(new Car("mo",3));
-        joinCars.add(new Car("jae",2));
-        progressRecord = new ProgressRecord(joinCars);
-        
-        List<Car> winner = List.of(new Car("crong",5));
-        
-        assertThat(progressRecord.findWinners()).isEqualTo(winner);
-    }
-    
-    @Test
     void showGameResultTest() {
         List<Car> joinCars = new ArrayList<>();
         joinCars.add(new Car("pobi"));
